~ completely refactor endpoint logic
